refactor: create abstract cloudevent package
