+++ cloudevents/abstract/__init__.py
@@ -11,3 +11,5 @@
 #    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 #    License for the specific language governing permissions and limitations
 #    under the License.
+
+from cloudevents.abstract.event import CloudEvent, AnyCloudEvent  # noqa
